Conflicts:
	common/src/main/java/com/scholarscore/models/Section.java
	common/src/test/java/com/scholarscore/models/GradeFormulaTest.java
	warehouse/src/main/java/com/scholarscore/api/controller/StudentAssignmentController.java
	warehouse/src/main/java/com/scholarscore/api/manager/StudentAssignmentManager.java
	warehouse/src/main/java/com/scholarscore/api/persistence/StudentAssignmentPersistence.java
	warehouse/src/main/java/com/scholarscore/api/persistence/goalCalculators/AssignmentGoalCalc.java
	warehouse/src/main/java/com/scholarscore/api/persistence/goalCalculators/AttendanceGoalCalc.java
	warehouse/src/main/java/com/scholarscore/api/persistence/mysql/jdbc/StudentAssignmentJdbc.java
	warehouse/src/test/java/com/scholarscore/api/util/SchoolDataFactory.java

+++ warehouse/src/main/java/com/scholarscore/api/persistence/goalCalculators/AssignmentGoalCalc.java
@@ -20,7 +20,7 @@ public class AssignmentGoalCalc implements GoalCalc<CalculatableAssignment> {
 
     public Double calculateGoal(CalculatableAssignment goal) {
         StudentAssignment goalAssignment = studentAssignmentPersistence.select(goal.getStudent().getId(), goal.getParentId());
-        Long awardedPoints = goalAssignment.getAwardedPoints();
+        Double awardedPoints = goalAssignment.getAwardedPoints();
                 if (null != awardedPoints) {
                     return awardedPoints.doubleValue();
                 } else {
